# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem ujian online (bank soal, jadwal, pengerjaan ujian per section, penilaian).",
        "title": "Ujian Backend API",
        "contact": {
            "name": "Tim Backend Ujian"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/mapel": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "List mata pelajaran",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah per halaman",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all mapel successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_mapel_dto.MapelListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "Buat mata pelajaran baru",
                "parameters": [
                    {
                        "description": "Data mapel",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_mapel_dto.CreateMapelRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create mapel successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_mapel_dto.MapelResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/mapel/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "Detail mata pelajaran",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID mapel (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get mapel successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_mapel_dto.MapelResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Mapel tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "Update mata pelajaran",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID mapel (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data mapel",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_mapel_dto.UpdateMapelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update mapel successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_mapel_dto.MapelResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validasi gagal",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "Hapus mata pelajaran (soft delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID mapel (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete mapel successfully",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Gagal menghapus",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/mapel/{id}/restore": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapel"
                ],
                "summary": "Restore mata pelajaran yang sudah dihapus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID mapel (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restore mapel successfully",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Gagal restore",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_internal_modules_mapel_dto.CreateMapelRequest": {
            "type": "object",
            "required": [
                "kode_mapel",
                "nama_mapel"
            ],
            "properties": {
                "deskripsi": {
                    "type": "string"
                },
                "kode_mapel": {
                    "type": "string",
                    "maxLength": 20
                },
                "nama_mapel": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_mapel_dto.MapelListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_mapel_dto.MapelResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_mapel_dto.MapelResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deskripsi": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kode_mapel": {
                    "type": "string"
                },
                "nama_mapel": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_mapel_dto.UpdateMapelRequest": {
            "type": "object",
            "required": [
                "kode_mapel",
                "nama_mapel"
            ],
            "properties": {
                "deskripsi": {
                    "type": "string"
                },
                "kode_mapel": {
                    "type": "string",
                    "maxLength": 20
                },
                "nama_mapel": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}